// src/interface/applicationlist/type.ts
export interface IApplicationList {
  key: string;
  no: string;
  state: any;
  name: string;
  code: string;
  company: string;
  apartment: string;
  job: string;
  apply: string;
  because: string;
  day: string;
  reason: string;
  from: string;
  to: string;
  rest: string;
  manager: string;
  dateApply: string;
  

  // ...
}